change balance to float and view price for service

+++ app/db/models/models.py
@@ -10,7 +10,7 @@ class User(Base):
     first_name = Column(String(256), nullable=False)
     promocode = Column(String(256), nullable=True)
     marketer = Column(Boolean, default=False, server_default="false", nullable=False)
-    balance = Column(BigInteger, nullable=True, default=0, server_default='0')
+    balance = Column(Float, nullable=True, default=0, server_default='0')
     registration_day = Column(Date, nullable=False, default=datetime.today().date, server_default=f"{datetime.today().strftime('%Y-%m-%d')}")
     is_payed_first_time = Column(Boolean, nullable=False, default=False, server_default='false')
     # promos = relationship("Promo", back_populates="user", lazy='joined', cascade='all, delete-orphan', uselist=True)
